Module opentelemetry::api::context::propagation::binary_propagator

Binary Propagator

BinaryFormat is a formatter to serialize and deserialize a value into a binary format.

BinaryFormat MUST expose the APIs that serializes values into bytes, and deserializes values from bytes.

Structs

BinaryPropagator
Extracts and injects SpanContexts from byte arrays.

Traits

BinaryFormat
Used to serialize and deserialize SpanContexts to and from a binary representation.