Trait opentelemetry::api::context::propagation::text_propagator::HttpTextFormat

pub trait HttpTextFormat: Debug {
    fn inject_context(&self, cx: &Context, carrier: &mut dyn Carrier);

    fn extract_with_context(
        &self, 
        cx: &Context, 
        carrier: &dyn Carrier
    ) -> Context;

    fn inject(&self, carrier: &mut dyn Carrier) { ... }

    fn extract(&self, carrier: &dyn Carrier) -> Context { ... }
}

Methods to inject and extract a value as text into carriers that travel in-band across process boundaries.

Required methods

fn inject_context(&self, cx: &Context, carrier: &mut dyn Carrier)

Properly encodes the values of the Context and injects them into the Carrier.

fn extract_with_context(&self, cx: &Context, carrier: &dyn Carrier) -> Context

Retrieves encoded data using the provided Carrier. If no data for this format was retrieved OR if the retrieved data is invalid, then the given Context is returned.

Loading content...
Provided methods

fn inject(&self, carrier: &mut dyn Carrier)

Properly encodes the values of the current Context and injects them into the Carrier.

fn extract(&self, carrier: &dyn Carrier) -> Context

Retrieves encoded data using the provided Carrier. If no data for this format was retrieved OR if the retrieved data is invalid, then the current Context is returned.

Loading content...