Struct opentelemetry::api::correlation::CorrelationContext

pub struct CorrelationContext { /* fields omitted */ }

A set of name/value pairs describing user-defined properties across systems.

Implementations

impl CorrelationContext

pub fn new() -> Self

Creates an empty CorrelationContext.

pub fn get<T: Into<Key>>(&self, key: T) -> Option<&Value>

Returns a reference to the value associated with a given name

Examples

use opentelemetry::api::{CorrelationContext, Value};

let mut cc = CorrelationContext::new();
let _ = cc.insert("my-name", "my-value");

assert_eq!(cc.get("my-name"), Some(&Value::String("my-value".to_string())))

pub fn insert<K, V>(&mut self, key: K, value: V) -> Option<Value> where
K: Into<Key>,
V: Into<Value>,

Inserts a name-value pair into the correlation context.

If the name was not present, [None] is returned. If the name was present, the value is updated, and the old value is returned.

Examples

use opentelemetry::api::{CorrelationContext, Value};

let mut cc = CorrelationContext::new();
let _ = cc.insert("my-name", "my-value");

assert_eq!(cc.get("my-name"), Some(&Value::String("my-value".to_string())))

pub fn remove<K: Into<Key>>(&mut self, key: K) -> Option<Value>

Removes a name from the correlation context, returning the value corresponding to the name if the pair was previously in the map.

pub fn len(&self) -> usize

Returns the number of attributes for this correlation context

pub fn is_empty(&self) -> bool

Returns true if the correlation context contains no items.

pub fn iter(&self) -> Iter

Gets an iterator over the correlation context items, sorted by name.