Module opentelemetry::api::distributed_context

OpenTelemetry Distributed Context API

OpenTelemetry uses Propagators to serialize and deserialize SpanContext into a binary or text format. Currently there are two types of propagators:

BinaryFormat which is used to serialize and deserialize a value into a binary representation.

HTTPTextFormat which is used to inject and extract a value as text into Carriers that travel in-band across process boundaries.

Modules

binary_propagator
Binary Propagator

http_b3_propagator
HTTP B3 Propagator

http_trace_context_propagator
HTTP Trace Context Propagator