Trait opentelemetry::api::metrics::counter::CounterHandle

pub trait CounterHandle<T>: InstrumentHandle where
    T: Into<MeasurementValue>, {
    fn add(&self, value: T) { ... }
}

CounterHandle is a handle for Counter instances.

It allows for repeated add calls for a pre-determined LabelSet.

Provided methods

fn add(&self, value: T)

Add works by calling the underlying record_one method available because this trait also implements InstrumentHandle.

Loading content...