Trait opentelemetry::api::metrics::gauge::Gauge

pub trait Gauge<T, LS>: Instrument<LS> where
    T: Into<MeasurementValue>,
    LS: LabelSet, {
    type Handle: GaugeHandle<T>;
    fn measurement(&self, value: T) -> Measurement<LS>;

    fn acquire_handle(&self, labels: &LS) -> Self::Handle;

    fn set(&self, value: T, label_set: &LS) { ... }
}

An interface for recording values where the metric cannot be expressed as a sum or because the measurement interval is arbitrary.

Associated Types

type Handle: GaugeHandle<T>

The handle type for the implementing Gauge.

Loading content...
Required methods

fn measurement(&self, value: T) -> Measurement<LS>

Creates a Measurement object to be used by a Meter when batch recording.

fn acquire_handle(&self, labels: &LS) -> Self::Handle

Creates a handle for this gauge. The labels should contain the keys and values for each key specified in the LabelSet.

If the labels do not contain a value for the key specified in the LabelSet, then the missing value will be treated as unspecified.

Loading content...
Provided methods

fn set(&self, value: T, label_set: &LS)

Assigns the passed value to the value of the gauge. The labels should contain the keys and values for each key specified in the LabelSet.

If the labels do not contain a value for the key specified in the LabelSet, then the missing value will be treated as unspecified.

Loading content...