Trait opentelemetry::api::metrics::gauge::GaugeHandle

pub trait GaugeHandle<T>: InstrumentHandle where
    T: Into<MeasurementValue>, {
    fn set(&self, value: T) { ... }
}

GaugeHandle is a handle for Gauge instances.

It allows for repeated set calls for a pre-determined LabelSet.

Provided methods

fn set(&self, value: T)

Set works by calling the underlying record_one method available because this trait also implements InstrumentHandle.

Loading content...