Module opentelemetry::api::metrics::measure

Metrics Measure Interface

Measures support record(value, label_set), signifying that events report individual measurements. This kind of metric should be used when the count or rate of events is meaningful and either:

The sum is of interest in addition to the count (rate)

Quantile information is of interest.

Measures are defined as with_absolute(true) by default, meaning that negative values are invalid. absolute = true measures are typically used to record absolute values such as durations and sizes.

When passing MetricOptions, measures can be declared as with_absolute(false) to indicate support for positive and negative values.

Traits

Measure
An interface for recording values where the count or rate of events is meaningful.

MeasureHandle
MeasureHandle is a handle for Measure instances.