Trait opentelemetry::api::trace::futures::Instrument

pub trait Instrument: Sized {
    fn instrument<S: Span>(self, span: S) -> Instrumented<Self, S> { ... }

    fn in_active_span<T: Tracer>(self, tracer: T) -> Instrumented<Self, T::Span> { ... }
}

Extension trait allowing futures, streams, sinks, and executors to be traced with a span.

Provided methods

Important traits for Instrumented<F, S>
impl<F: Future, S: Span> Future for Instrumented<F, S> type Output = F::Output;

fn instrument<S: Span>(self, span: S) -> Instrumented<Self, S>

Traces this type with the provided Span, returning a Instrumented wrapper.

Important traits for Instrumented<F, S>
impl<F: Future, S: Span> Future for Instrumented<F, S> type Output = F::Output;

fn in_active_span<T: Tracer>(self, tracer: T) -> Instrumented<Self, T::Span>

Traces this type with the provided Tracer's active span, returning a Instrumented wrapper.

Loading content...