Module opentelemetry::api::trace::provider

OpenTelemetry Tracer Provider Interface

Obtaining a Tracer

New Tracer instances can be created via a Provider and its get_tracer method. This method expects an Into argument:

name (required): This name must identify the instrumentation library (also referred to as integration, e.g. io.opentelemetry.contrib.mongodb) and not the instrumented library. In case an invalid name (empty string) is specified, a working default Tracer implementation as a fallback is returned rather than returning None or throwing an exception. A library, implementing the OpenTelemetry API may also ignore this name and return a default instance for all calls, if it does not support "named" functionality (e.g. an implementation which is not even observability-related). A Provider could also return a no-op Tracer here if application owners configure the SDK to suppress telemetry produced by this library.

Implementations might require the user to specify configuration properties at Provider creation time, or rely on external configuration.

Traits

Provider
An interface to create Tracer instances.