Trait opentelemetry::api::trace::span_processor::SpanProcessor

pub trait SpanProcessor: Send + Sync + Debug {
    fn on_start(&self, span: Arc<SpanData>);

    fn on_end(&self, span: Arc<SpanData>);

    fn shutdown(&self);
}

SpanProcessors allow finished spans to be processed.

Required methods

fn on_start(&self, span: Arc<SpanData>)

on_start method is invoked when a Span is started.

fn on_end(&self, span: Arc<SpanData>)

on_end method is invoked when a Span is ended.

fn shutdown(&self)

Shutdown is invoked when SDK shuts down. Use this call to cleanup any processor data. No calls to on_start and on_end method is invoked after shutdown call is made.

Loading content...