Module opentelemetry::api::trace::trace_context_propagator

Trace Context Propagator

The traceparent header represents the incoming request in a tracing system in a common format, understood by all vendors. Here’s an example of a traceparent header.

traceparent: 00-0af7651916cd43dd8448eb211c80319c-b7ad6b7169203331-01

The traceparent HTTP header field identifies the incoming request in a tracing system. It has four fields:

version

trace-id

parent-id

trace-flags

See the w3c trace-context docs for more details.

Structs

TraceContextPropagator
Extracts and injects SpanContexts into Carriers using the trace-context format.