Module opentelemetry::exporter::trace::stdout

Stdout Span Exporter

The stdout SpanExporter writes debug printed Spans to its configured Write instance. By default it will write to Stdout.

Examples

use opentelemetry::exporter::trace::stdout;
use opentelemetry::{sdk, global};

// Create a new stdout exporter that writes pretty printed span output
let exporter = stdout::Builder::default().with_pretty_print(true).init();
let provider = sdk::Provider::builder()
    .with_simple_exporter(exporter)
    .build();
global::set_provider(provider);

Structs

Builder
Builder

Exporter
A SpanExporter that writes to Stdout or other configured Write.