Function opentelemetry::global::tracer

pub fn tracer(name: &'static str) -> BoxedTracer

Creates a named instance of Tracer via the configured GlobalTracerProvider.

If the name is an empty string, the provider will use a default name.

This is a more convenient way of expressing global::tracer_provider().get_tracer(name, None).