Function opentelemetry::global::tracer_with_version

pub fn tracer_with_version(
    name: &'static str, 
    version: &'static str
) -> BoxedTracer

Creates a named instance of Tracer with version info via the configured GlobalTracerProvider

If the name is an empty string, the provider will use a default name. If the version is an empty string, it will be used as part of instrumentation library information.