Trait opentelemetry::global::GenericTracerProvider

pub trait GenericTracerProvider: Debug + 'static {
    fn get_tracer_boxed(
        &self, 
        name: &'static str, 
        version: Option<&'static str>
    ) -> Box<dyn GenericTracer + Send + Sync>;
}

Allows a specific TracerProvider to be used generically by the GlobalTracerProvider by mirroring the interface and boxing the return types.

Required methods

fn get_tracer_boxed(
&self,
name: &'static str,
version: Option<&'static str>
) -> Box<dyn GenericTracer + Send + Sync>

Creates a named tracer instance that is a trait object through the underlying TracerProvider.

Loading content...