opentelemetry::global

Trait ObjectSafeTracerProvider

pub trait ObjectSafeTracerProvider {
    // Required method
    fn boxed_tracer(
        &self,
        scope: InstrumentationScope,
    ) -> Box<dyn ObjectSafeTracer + Send + Sync>;
}

Allows a specific TracerProvider to be used generically by the GlobalTracerProvider by mirroring the interface and boxing the return types.

Required Methods

fn boxed_tracer( &self, scope: InstrumentationScope, ) -> Box<dyn ObjectSafeTracer + Send + Sync>

Creates a versioned named tracer instance that is a trait object through the underlying TracerProvider.