Struct opentelemetry::metrics::InstrumentBuilder

pub struct InstrumentBuilder<'a, T> { /* private fields */ }

Configuration for building a sync instrument.

Implementations

impl<'a, T> InstrumentBuilder<'a, T>where T: TryFrom<InstrumentBuilder<'a, T>, Error = MetricsError>,

pub fn with_description<S>(self, description: S) -> InstrumentBuilder<'a, T>where S: Into<Cow<'static, str>>,

Set the description for this instrument

pub fn with_unit(self, unit: Unit) -> InstrumentBuilder<'a, T>

Set the unit for this instrument.

Unit is case sensitive(kb is not the same as kB).

Unit must be:

ASCII string

No longer than 63 characters

pub fn try_init(self) -> Result<T, MetricsError>

Validate the instrument configuration and creates a new instrument.

pub fn init(self) -> T

Creates a new instrument.

This method reports configuration errors but still returns the instrument.

Panics

Panics if the instrument cannot be created. Use try_init if you want to handle errors.