opentelemetry::metrics
Trait CallbackRegistration

pub trait CallbackRegistration: Send + Sync {
    // Required method
    fn unregister(&mut self) -> Result<(), MetricsError>;
}

Available on crate feature metrics only.

A token representing the unique registration of a callback for a set of instruments with a Meter.

Required Methods

fn unregister(&mut self) -> Result<(), MetricsError>

Removes the callback registration from its associated Meter.