opentelemetry::metrics
Trait InstrumentProvider

pub trait InstrumentProvider {

    // Required method
    fn register_callback(
        &self,
        callback: Box<dyn Fn(&Context) + Send + Sync>,
    ) -> Result<(), MetricsError>;

    // Provided methods
    fn u64_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<Counter<u64>, MetricsError> { ... }
    fn f64_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<Counter<f64>, MetricsError> { ... }
    fn u64_observable_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableCounter<u64>, MetricsError> { ... }
    fn f64_observable_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableCounter<f64>, MetricsError> { ... }
    fn i64_up_down_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<UpDownCounter<i64>, MetricsError> { ... }
    fn f64_up_down_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<UpDownCounter<f64>, MetricsError> { ... }
    fn i64_observable_up_down_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableUpDownCounter<i64>, MetricsError> { ... }
    fn f64_observable_up_down_counter(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableUpDownCounter<f64>, MetricsError> { ... }
    fn u64_observable_gauge(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableGauge<u64>, MetricsError> { ... }
    fn i64_observable_gauge(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableGauge<i64>, MetricsError> { ... }
    fn f64_observable_gauge(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<ObservableGauge<f64>, MetricsError> { ... }
    fn f64_histogram(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<Histogram<f64>, MetricsError> { ... }
    fn u64_histogram(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<Histogram<u64>, MetricsError> { ... }
    fn i64_histogram(
        &self,
        _name: String,
        _description: Option<String>,
        _unit: Option<Unit>,
    ) -> Result<Histogram<i64>, MetricsError> { ... }

}

Available on crate feature metrics only.

SDK implemented trait for creating instruments

Required Methods

fn register_callback( &self, callback: Box<dyn Fn(&Context) + Send + Sync>, ) -> Result<(), MetricsError>

Captures the function that will be called during data collection.

It is only valid to call observe within the scope of the passed function.

Provided Methods

fn u64_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<Counter<u64>, MetricsError>

creates an instrument for recording increasing values.

fn f64_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<Counter<f64>, MetricsError>

creates an instrument for recording increasing values.

fn u64_observable_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableCounter<u64>, MetricsError>

creates an instrument for recording increasing values via callback.

fn f64_observable_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableCounter<f64>, MetricsError>

creates an instrument for recording increasing values via callback.

fn i64_up_down_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<UpDownCounter<i64>, MetricsError>

creates an instrument for recording changes of a value.

fn f64_up_down_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<UpDownCounter<f64>, MetricsError>

creates an instrument for recording changes of a value.

fn i64_observable_up_down_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableUpDownCounter<i64>, MetricsError>

creates an instrument for recording changes of a value.

fn f64_observable_up_down_counter( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableUpDownCounter<f64>, MetricsError>

creates an instrument for recording changes of a value via callback.

fn u64_observable_gauge( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableGauge<u64>, MetricsError>

creates an instrument for recording the current value via callback.

fn i64_observable_gauge( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableGauge<i64>, MetricsError>

creates an instrument for recording the current value via callback.

fn f64_observable_gauge( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<ObservableGauge<f64>, MetricsError>

creates an instrument for recording the current value via callback.

fn f64_histogram( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<Histogram<f64>, MetricsError>

creates an instrument for recording a distribution of values.

fn u64_histogram( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<Histogram<u64>, MetricsError>

creates an instrument for recording a distribution of values.

fn i64_histogram( &self, _name: String, _description: Option<String>, _unit: Option<Unit>, ) -> Result<Histogram<i64>, MetricsError>

creates an instrument for recording a distribution of values.