opentelemetry::metrics
Trait MeterProvider

pub trait MeterProvider {
    // Required method
    fn versioned_meter(
        &self,
        name: impl Into<Cow<'static, str>>,
        version: Option<impl Into<Cow<'static, str>>>,
        schema_url: Option<impl Into<Cow<'static, str>>>,
        attributes: Option<Vec<KeyValue>>,
    ) -> Meter;

    // Provided method
    fn meter(&self, name: impl Into<Cow<'static, str>>) -> Meter { ... }
}

Available on crate feature metrics only.

Provides access to named Meter instances, for instrumenting an application or crate.

Required Methods

fn versioned_meter( &self, name: impl Into<Cow<'static, str>>, version: Option<impl Into<Cow<'static, str>>>, schema_url: Option<impl Into<Cow<'static, str>>>, attributes: Option<Vec<KeyValue>>, ) -> Meter

Creates an implementation of the Meter interface.

The instrumentation name must be the name of the library providing instrumentation. This name may be the same as the instrumented code only if that code provides built-in instrumentation. If the instrumentation name is empty, then a implementation defined default name will be used instead.

Provided Methods

fn meter(&self, name: impl Into<Cow<'static, str>>) -> Meter

Returns a new Meter with the provided name and default configuration.

A Meter should be scoped at most to a single application or crate. The name needs to be unique so it does not collide with other names used by an application, nor other applications.

If the name is empty, then an implementation defined default name will be used instead.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.