Module opentelemetry::sdk

OpenTelemetry SDK

This SDK provides an opinionated reference implementation of the OpenTelemetry API. The SDK implements the specifics of deciding which data to collect through Samplers, and facilitates the delivery of telemetry data to storage systems through Exporters. These can be configured on Tracer and Meter creation.

Re-exports

pub use metrics::LabelSet;

pub use metrics::Meter;

pub use resource::Resource;

pub use trace::config::Config;

pub use trace::evicted_hash_map::EvictedHashMap;

pub use trace::evicted_queue::EvictedQueue;

pub use trace::id_generator::IdGenerator;

pub use trace::provider::Provider;

pub use trace::sampler::Sampler;

pub use trace::span::Span;

pub use trace::span_processor::BatchSpanProcessor;

pub use trace::span_processor::SimpleSpanProcessor;

pub use trace::tracer::Tracer;

Modules

metrics
OpenTelemetry Metrics SDK

resource
Resource

trace
OpenTelemetry Trace SDK