Struct opentelemetry::sdk::export::metrics::aggregation::AggregationKind

pub struct AggregationKind(_);

Available on crate feature metrics only.

A short name for the Aggregator that produces an Aggregation.

Kind is a string to allow user-defined Aggregators.

When deciding how to handle an Aggregation, Exporters are encouraged to decide based on conversion to the above interfaces based on strength, not on Kind value, when deciding how to expose metric data. This enables user-supplied Aggregators to replace builtin Aggregators.

For example, test for a Histogram before testing for a Sum, and so on.

Implementations

impl AggregationKind

pub const SUM: AggregationKind = AggregationKind("SUM")

Aggregations that return an aggregated sum.

pub const HISTOGRAM: AggregationKind = AggregationKind("HISTOGRAM")

Aggregations that return a distribution

pub const LAST_VALUE: AggregationKind = AggregationKind("LAST_VALUE")

Aggregations that return only the latest value.

pub const fn new(name: &'static str) -> AggregationKind

Create a new custom aggregation kind