opentelemetry::sdk::export::metrics::aggregation
Trait TemporalitySelector

pub trait TemporalitySelector {
    // Required method
    fn temporality_for(
        &self,
        descriptor: &Descriptor,
        kind: &AggregationKind,
    ) -> Temporality;
}

Available on crate feature metrics only.

TemporalitySelector is a sub-interface of Exporter used to indicate whether the Processor should compute Delta or Cumulative Aggregations.

Required Methods

fn temporality_for( &self, descriptor: &Descriptor, kind: &AggregationKind, ) -> Temporality

TemporalityFor should return the correct Temporality that should be used when exporting data for the given metric instrument and Aggregator kind.