Module opentelemetry::sdk::export::trace

Trace exporters

Modules

stdout

Stdout Span Exporter

Structs

SpanData

SpanData contains all the information collected by a Span and can be used by exporters as a standard input.

Traits

SpanExporter

SpanExporter defines the interface that protocol-specific exporters must implement so that they can be plugged into OpenTelemetry SDK and support sending of telemetry data.

Type Definitions

ExportResult

Describes the result of an export.