Module opentelemetry::sdk::export::trace::stdout

Stdout Span Exporter

The stdout SpanExporter writes debug printed Spans to its configured Write instance. By default it will write to Stdout.

Examples

use opentelemetry::trace::Tracer;
use opentelemetry::sdk::export::trace::stdout;

fn main() {
    let (tracer, _uninstall) = stdout::new_pipeline()
        .with_pretty_print(true)
        .install();

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });
}

Structs

Exporter

A SpanExporter that writes to Stdout or other configured Write.

PipelineBuilder

Pipeline builder

Uninstall

Uninstalls the stdout pipeline on drop.

Functions

new_pipeline

Create a new stdout exporter pipeline builder.