opentelemetry::sdk::metrics
Module aggregators

Available on crate feature metrics only.

Metric Aggregators

Structs

HistogramAggregator

This aggregator observes events and counts them in pre-determined buckets. It also calculates the sum and count of all events.

LastValueAggregator

Aggregates last value events.

SumAggregator

An aggregator for counter events.

Traits

Aggregator

Aggregator implements a specific aggregation behavior, i.e., a behavior to track a sequence of updates to an instrument. Sum-only instruments commonly use a simple Sum aggregator, but for the distribution instruments (Histogram, ValueObserver) there are a number of possible aggregators with different cost and accuracy tradeoffs.

Functions

histogram

Create a new histogram for the given descriptor with the given boundaries

last_value

Create a new LastValueAggregator

range_test

RangeTest is a common routine for testing for valid input values. This rejects NaN values. This rejects negative values when the metric instrument does not support negative values, including monotonic counter metrics and absolute Histogram metrics.

sum

Create a new sum aggregator.