Struct opentelemetry::sdk::metrics::Meter

pub struct Meter { /* fields omitted */ }

Meter implementation to create manage metric instruments and record batch measurements

Implementations

impl Meter

pub fn new(component: &'static str) -> Self

Create a new Meter instance with a component name and empty registry.