Struct opentelemetry::sdk::metrics::Stream

pub struct Stream {
    pub name: Cow<'static, str>,
    pub description: Cow<'static, str>,
    pub unit: Unit,
    pub aggregation: Option<Aggregation>,
    pub allowed_attribute_keys: Option<Arc<HashSet<Key, RandomState>, Global>>,
}

Describes the stream of data an instrument produces.

Example

Streams can be used as masks in views.

use opentelemetry_sdk::metrics::{new_view, Aggregation, Instrument, Stream};

let criteria = Instrument::new().name("counter_*");
let mask = Stream::new().aggregation(Aggregation::Sum);

let view = new_view(criteria, mask);

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Cow<'static, str>

The human-readable identifier of the stream.

description: Cow<'static, str>

Describes the purpose of the data.

unit: Unit

the unit of measurement recorded.

aggregation: Option<Aggregation>

Aggregation the stream uses for an instrument.

allowed_attribute_keys: Option<Arc<HashSet<Key, RandomState>, Global>>

An allow-list of attribute keys that will be preserved for the stream.

Any attribute recorded for the stream with a key not in this set will be dropped. If the set is empty, all attributes will be dropped, if None all attributes will be kept.

Implementations

impl Stream

pub fn new() -> Stream

Create a new stream with empty values.

pub fn name(self, name: impl Into<Cow<'static, str>>) -> Stream

Set the stream name.

pub fn description(self, description: impl Into<Cow<'static, str>>) -> Stream

Set the stream description.

pub fn unit(self, unit: Unit) -> Stream

Set the stream unit.

pub fn aggregation(self, aggregation: Aggregation) -> Stream

Set the stream aggregation.

pub fn allowed_attribute_keys( self, attribute_keys: impl IntoIterator<Item = Key> ) -> Stream

Set the stream allowed attribute keys.

Any attribute recorded for the stream with a key not in this set will be dropped. If this set is empty all attributes will be dropped.