opentelemetry::sdk
Module resource

Representations of entities producing telemetry.

A Resource is an immutable representation of the entity producing telemetry as attributes. For example, a process producing telemetry that is running in a container on Kubernetes has a Pod name, it is in a namespace and possibly is part of a Deployment which also has a name. All three of these attributes can be included in the Resource. Note that there are certain “standard attributes” that have prescribed meanings.

Resource detectors

ResourceDetectors are used to detect resource from runtime or environmental variables. The following are provided by default with this SDK.

EnvResourceDetector - detect resource from environmental variables.

OsResourceDetector - detect OS from runtime.

ProcessResourceDetector - detect process information.

TelemetryResourceDetector - detect telemetry SDK’s information.

Structs

EnvResourceDetector

Resource detector implements ResourceDetector and is used to extract general SDK configuration from environment.

IntoIter

An owned iterator over the entries of a Resource.

Iter

An iterator over the entries of a Resource.

OsResourceDetector

Detect runtime operating system information.

ProcessResourceDetector

Detect process information.

Resource

An immutable representation of the entity producing telemetry as attributes.

SdkProvidedResourceDetector

There are attributes which MUST be provided by the SDK as specified in the Resource SDK specification. This detector detects those attributes and if the attribute cannot be detected, it uses the default value.

TelemetryResourceDetector

Detect the telemetry SDK information used to capture data recorded by the instrumentation libraries.

Traits

ResourceDetector

ResourceDetector detects OpenTelemetry resource information