Module opentelemetry::sdk::trace::provider

Trace Provider SDK

Tracer Creation

New Tracer instances are always created through a Provider.

All configuration objects and extension points (span processors, propagators) are provided by the Provider. Tracer instances do not duplicate this data to avoid that different Tracer instances of the Provider have different versions of these data.

Structs

Builder
Builder for provider attributes.

Provider
Creator and registry of named Tracer instances.