Struct opentelemetry::sdk::trace::span_processor::BatchSpanProcessorWorker

pub struct BatchSpanProcessorWorker { /* fields omitted */ }

A worker process that batches and processes spans as they are reported.

This process is implemented as a [Future] that returns when the accompanying [BatchSpanProcessor] is shut down, and allows systems like [tokio] and [async-std] to process the work in the background without requiring dedicated system threads.