Struct opentelemetry::sdk::trace::SimpleSpanProcessor

pub struct SimpleSpanProcessor { /* fields omitted */ }

A SpanProcessor that exports synchronously when spans are finished.

Examples

Note that the simple processor exports synchronously every time a span is ended. If you find this limiting, consider the batch processor instead.

use opentelemetry::{trace as apitrace, sdk::trace as sdktrace, global};

// Configure your preferred exporter
let exporter = apitrace::NoopSpanExporter::new();

// Then use the `with_simple_exporter` method to have the provider export when spans finish.
let provider = sdktrace::TracerProvider::builder()
    .with_simple_exporter(exporter)
    .build();

let previous_provider = global::set_tracer_provider(provider);