Enum opentelemetry::trace::TraceError

#[non_exhaustive]
pub enum TraceError {
    ExportFailed(Box<dyn ExportError>),
    ExportTimedOut(Duration),
    Other(Box<dyn Error + Send + Sync + 'static>),
}

Errors returned by the trace API.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ExportFailed(Box<dyn ExportError>)

Export failed with the error returned by the exporter

ExportTimedOut(Duration)

Export failed to finish after certain period and processor stopped the export.

Other(Box<dyn Error + Send + Sync + 'static>)

Other errors propagated from trace SDK that werenโt covered above