Struct opentelemetry::trace::SpanRef

pub struct SpanRef<'a>(_);

A reference to the currently active span in this context.

Implementations

impl SpanRef<'_>

pub fn add_event<T>(&self, name: T, attributes: Vec<KeyValue>) where
T: Into<Cow<'static, str>>,

An API to record events in the context of a given Span.

pub fn record_exception(&self, err: &dyn Error)

Convenience method to record an exception/error as an Event

pub fn record_exception_with_stacktrace<T>(
&self,
err: &dyn Error,
stacktrace: T
) where
T: Into<Cow<'static, str>>,

Convenience method to record a exception/error as an Event with custom stacktrace

pub fn add_event_with_timestamp<T>(
&self,
name: T,
timestamp: SystemTime,
attributes: Vec<KeyValue>
) where
T: Into<Cow<'static, str>>,

An API to record events at a specific time in the context of a given Span.

pub fn span_context(&self) -> &SpanContext

Returns the SpanContext for the given Span.

pub fn is_recording(&self) -> bool

Returns true if this Span is recording information like events with the add_event operation, attributes using set_attributes, status with set_status, etc.

pub fn set_attribute(&self, attribute: KeyValue)

An API to set a single Attribute where the attribute properties are passed as arguments. To avoid extra allocations some implementations may offer a separate API for each of the possible value types.

pub fn set_status(&self, code: StatusCode, message: String)

Sets the status of the Span. If used, this will override the default Span status, which is Unset. message MUST be ignored when the status is OK or Unset

pub fn update_name<T>(&self, new_name: String) where
T: Into<Cow<'static, str>>,

Updates the Span’s name. After this update, any sampling behavior based on the name will depend on the implementation.

pub fn end(&self)

Finishes the Span.

pub fn end_with_timestamp(&self, timestamp: SystemTime)

Finishes the Span with given timestamp