Trait opentelemetry::trace::FutureExt

pub trait FutureExt: Sized {
    // Provided methods
    fn with_context(self, otel_cx: Context) -> WithContext<Self>  { ... }
    fn with_current_context(self) -> WithContext<Self>  { ... }
}

Extension trait allowing futures, streams, and sinks to be traced with a span.

Provided Methods

fn with_context(self, otel_cx: Context) -> WithContext<Self>

Attaches the provided Context to this type, returning a WithContext wrapper.

When the wrapped type is a future, stream, or sink, the attached context will be set as current while it is being polled.

fn with_current_context(self) -> WithContext<Self>

Attaches the current Context to this type, returning a WithContext wrapper.

When the wrapped type is a future, stream, or sink, the attached context will be set as the default while it is being polled.

Object Safety

This trait is not object safe.