Trait opentelemetry::api::trace::tracer::Tracer

pub trait Tracer: Send + Sync {
    type Span: Span;
    fn invalid(&self) -> Self::Span;

    fn start(&self, name: &str, parent_span: Option<SpanContext>) -> Self::Span;

    fn get_active_span(&self) -> Self::Span;

    fn mark_span_as_active(&self, span: &Self::Span);

    fn mark_span_as_inactive(&self, span_id: u64);
}

Interface for constructing Spans.

Associated Types

type Span: Span

The Span type used by this Tracer.

Loading content...
Required methods

fn invalid(&self) -> Self::Span

Returns a span with an invalid SpanContext. Used by functions that need to return a default span like get_active_span if no span is present.

fn start(&self, name: &str, parent_span: Option<SpanContext>) -> Self::Span

Starts a new Span.

By default the currently active Span is set as the new Span's parent. The Tracer MAY provide other default options for newly created Spans.

Span creation MUST NOT set the newly created Span as the currently active Span by default, but this functionality MAY be offered additionally as a separate operation.

Each span has zero or one parent spans and zero or more child spans, which represent causally related operations. A tree of related spans comprises a trace. A span is said to be a root span if it does not have a parent. Each trace includes a single root span, which is the shared ancestor of all other spans in the trace. Implementations MUST provide an option to create a Span as a root span, and MUST generate a new TraceId for each root span created. For a Span with a parent, the TraceId MUST be the same as the parent. Also, the child span MUST inherit all TraceState values of its parent by default.

A Span is said to have a remote parent if it is the child of a Span created in another process. Each propagators' deserialization must set is_remote to true on a parent SpanContext so Span creation knows if the parent is remote.

fn get_active_span(&self) -> Self::Span

Returns the current active span.

When getting the current Span, the Tracer MUST return a placeholder Span with an invalid SpanContext if there is no currently active Span.

fn mark_span_as_active(&self, span: &Self::Span)

Mark a given Span as active.

The Tracer MUST provide a way to update its active Span, and MAY provide convenience methods to manage a Span's lifetime and the scope in which a Span is active. When an active Span is made inactive, the previously-active Span SHOULD be made active. A Span maybe finished (i.e. have a non-null end time) but still be active. A Span may be active on one thread after it has been made inactive on another.

NOTE: The mark_span_as_active/mark_span_as_inactive functions MUST be used together or you can end up retaining references to the currently active Span. If you do not want to manage active state of Spans manually, use the with_span API defined for all Tracers via TracerGenerics

fn mark_span_as_inactive(&self, span_id: u64)

Remove span from active span

When an active Span is made inactive, the previously-active Span SHOULD be made active. A Span maybe finished (i.e. have a non-null end time) but still be active. A Span may be active on one thread after it has been made inactive on another.

NOTE: The mark_span_as_active/mark_span_as_inactive functions MUST be used together or you can end up retaining references to the currently active Span. If you do not want to manage active state of Spans manually, use the with_span API defined for all Tracers via TracerGenerics

Loading content...