Struct opentelemetry::baggage::Baggage

pub struct Baggage { /* fields omitted */ }

A set of name-value pairs describing user-defined properties.

Baggage Names

ASCII strings according to the token format, defined in RFC2616, Section 2.2

Baggage Values

URL encoded UTF-8 strings.

Baggage Value Metadata

Additional metadata can be added to values in the form of a property set, represented as semi-colon ; delimited list of names and/or name-value pairs, e.g. ;k1=v1;k2;k3=v3.

Limits

Maximum number of name-value pairs: 180.

Maximum number of bytes per a single name-value pair: 4096.

Maximum total length of all name-value pairs: 8192.

Implementations

impl Baggage

pub fn new() -> Self

Creates an empty Baggage.

pub fn get<T: Into<Key>>(&self, key: T) -> Option<&Value>

Returns a reference to the value associated with a given name

Examples

use opentelemetry::{baggage::Baggage, Value};

let mut cc = Baggage::new();
let _ = cc.insert("my-name", "my-value");

assert_eq!(cc.get("my-name"), Some(&Value::from("my-value")))

pub fn get_with_metadata<T: Into<Key>>(
&self,
key: T
) -> Option<&(Value, BaggageMetadata)>

Returns a reference to the value and metadata associated with a given name

Examples

use opentelemetry::{baggage::{Baggage, BaggageMetadata}, Value};

let mut cc = Baggage::new();
let _ = cc.insert("my-name", "my-value");

// By default, the metadata is empty
assert_eq!(cc.get_with_metadata("my-name"), Some(&(Value::from("my-value"), BaggageMetadata::from(""))))

pub fn insert<K, V>(&mut self, key: K, value: V) -> Option<Value> where
K: Into<Key>,
V: Into<Value>,

Inserts a name-value pair into the baggage.

If the name was not present, None is returned. If the name was present, the value is updated, and the old value is returned.

Examples

use opentelemetry::{baggage::Baggage, Value};

let mut cc = Baggage::new();
let _ = cc.insert("my-name", "my-value");

assert_eq!(cc.get("my-name"), Some(&Value::from("my-value")))

pub fn insert_with_metadata<K, V, S>(
&mut self,
key: K,
value: V,
metadata: S
) -> Option<(Value, BaggageMetadata)> where
K: Into<Key>,
V: Into<Value>,
S: Into<BaggageMetadata>,

Inserts a name-value pair into the baggage.

Same with insert, if the name was not present, None will be returned. If the name is present, the old value and metadata will be returned.

Examples

use opentelemetry::{baggage::{Baggage, BaggageMetadata}, Value};

let mut cc = Baggage::new();
let _ = cc.insert_with_metadata("my-name", "my-value", "test");

assert_eq!(cc.get_with_metadata("my-name"), Some(&(Value::from("my-value"), BaggageMetadata::from("test"))))

pub fn remove<K: Into<Key>>(
&mut self,
key: K
) -> Option<(Value, BaggageMetadata)>

Removes a name from the baggage, returning the value corresponding to the name if the pair was previously in the map.

pub fn len(&self) -> usize

Returns the number of attributes for this baggage

pub fn is_empty(&self) -> bool

Returns true if the baggage contains no items.

pub fn iter(&self) -> Iter<'_>

Gets an iterator over the baggage items, sorted by name.