Module opentelemetry::global

OpenTelemetry global Tracer and Meter singletons.

Structs

BoxedSpan
Boxed span wraps a generic trait object so that BoxedTracers can return whichever type of span they were configured to use.

BoxedTracer
BoxedTracer is an instance of a generic tracer that can be returned by the global provider to represent.

GlobalProvider
GlobalProvider maintains a global singleton that allows any thread to access the same generic Provider implementation.

Traits

GenericProvider
GenericProvider allows GlobalProvider's to contain and use a Provider trait object.

GenericTracer
GenericTracer allows BoxedTracer's to contain and use a Tracer trait object.

Functions

global_meter
Returns NoopMeter for now

set_provider
Assigns the global Tracer

trace_provider
Returns a reference to the global Provider