Function opentelemetry::global::set_tracer_provider

pub fn set_tracer_provider<P, T, S>(new_provider: P) -> GlobalTracerProviderwhere
    S: Span + Send + Sync + 'static,
    T: Tracer<Span = S> + Send + Sync + 'static,
    P: TracerProvider<Tracer = T> + Send + Sync + 'static,

Sets the given TracerProvider instance as the current global provider.

It returns the TracerProvider instance that was previously mounted as global provider (e.g. NoopTracerProvider if a provider had not been set before).