Trait opentelemetry::global::GenericTracer

pub trait GenericTracer: Send + Sync {
    fn invalid_boxed(&self) -> Box<dyn Span>;

    fn start_boxed(
        &self, 
        name: &str, 
        parent: Option<SpanContext>
    ) -> Box<dyn Span>;

    fn get_active_span_boxed(&self) -> Box<dyn Span>;

    fn mark_span_as_active_boxed(&self, span: &dyn Span);

    fn mark_span_as_inactive_boxed(&self, span_id: u64);
}

GenericTracer allows BoxedTracer's to contain and use a Tracer trait object.

Required methods

fn invalid_boxed(&self) -> Box<dyn Span>

Create a new invalid span for use in cases where there are no active spans.

fn start_boxed(&self, name: &str, parent: Option<SpanContext>) -> Box<dyn Span>

Returns a trait object so the underlying implementation can be swapped out at runtime.

fn get_active_span_boxed(&self) -> Box<dyn Span>

Returns the currently active span as a BoxedSpan

fn mark_span_as_active_boxed(&self, span: &dyn Span)

Returns the currently active span as a BoxedSpan

fn mark_span_as_inactive_boxed(&self, span_id: u64)

Marks the current span as inactive

Loading content...