Trait opentelemetry::global::ObjectSafeTracer

pub trait ObjectSafeTracer {
    fn start_with_context_boxed(
        &self, 
        name: Cow<'static, str>, 
        parent_cx: &Context
    ) -> Box<dyn ObjectSafeSpan + Send + Sync>;
    fn build_with_context_boxed(
        &self, 
        builder: SpanBuilder, 
        parent_cx: &Context
    ) -> Box<dyn ObjectSafeSpan + Send + Sync>;
}

Allows a specific Tracer to be used generically by BoxedTracer instances by mirroring the interface and boxing the return types.

Required Methods

fn start_with_context_boxed(
&self,
name: Cow<'static, str>,
parent_cx: &Context
) -> Box<dyn ObjectSafeSpan + Send + Sync>

Returns a trait object so the underlying implementation can be swapped out at runtime.

fn build_with_context_boxed(
&self,
builder: SpanBuilder,
parent_cx: &Context
) -> Box<dyn ObjectSafeSpan + Send + Sync>

Returns a trait object so the underlying implementation can be swapped out at runtime.