Trait opentelemetry::global::ObjectSafeTracerProvider

pub trait ObjectSafeTracerProvider {
    fn versioned_tracer_boxed(
        &self, 
        name: Cow<'static, str>, 
        version: Option<&'static str>, 
        schema_url: Option<&'static str>
    ) -> Box<dyn ObjectSafeTracer + Send + Sync>;
    fn force_flush(&self) -> Vec<TraceResult<()>>;
}

Allows a specific TracerProvider to be used generically by the GlobalTracerProvider by mirroring the interface and boxing the return types.

Required Methods

fn versioned_tracer_boxed(
&self,
name: Cow<'static, str>,
version: Option<&'static str>,
schema_url: Option<&'static str>
) -> Box<dyn ObjectSafeTracer + Send + Sync>

Creates a versioned named tracer instance that is a trait object through the underlying TracerProvider.

fn force_flush(&self) -> Vec<TraceResult<()>>

Force flush all remaining spans in span processors and return results.