Module opentelemetry::metrics

OpenTelemetry Metrics API

Modules

noop

No-op OpenTelemetry Metrics Implementation

Structs

Counter

An instrument that records increasing values.

Histogram

An instrument that records a distribution of values.

InstrumentBuilder

Configuration for building an instrument.

Meter

Provides access to instrument instances for recording metrics.

ObservableCounter

An async instrument that records increasing values.

ObservableGauge

An instrument that records independent readings.

ObservableUpDownCounter

An async instrument that records increasing or decreasing values.

Unit

Units denote underlying data units tracked by Meters.

UpDownCounter

An instrument that records increasing or decreasing values.

Enums

MetricsError

Errors returned by the metrics API.

Traits

AsyncCounter

An SDK implemented async instrument that records increasing values.

AsyncGauge

An SDK implemented instrument that records independent readings.

AsyncUpDownCounter

An SDK implemented async instrument that records increasing or decreasing values.

InstrumentProvider

SDK implemented trait for creating instruments

MeterProvider

Returns named meter instances

SyncCounter

An SDK implemented instrument that records increasing values.

SyncHistogram

An SDK implemented instrument that records a distribution of values.

SyncUpDownCounter

An SDK implemented instrument that records increasing or decreasing values.

Type Definitions

Result

A specialized Result type for metric operations.