Module opentelemetry::metrics::noop

No-op OpenTelemetry Metrics Implementation

This implementation is returned as the global Meter if no Meter has been set. It is also useful for testing purposes as it is intended to have minimal resource utilization and runtime impact.

Structs

NoopAsyncInstrument

A no-op async instrument.

NoopMeterCore

A no-op instance of a Meter

NoopMeterProvider

A no-op instance of a MetricProvider

NoopRegistration

A no-op instance of a callback CallbackRegistration.

NoopSyncInstrument

A no-op sync instrument