Struct opentelemetry::metrics::AsyncInstrumentBuilder

pub struct AsyncInstrumentBuilder<'a, I, M>
where
    I: AsyncInstrument<M>,
{ /* private fields */ }

Configuration for building an async instrument.

Implementations

impl<'a, I, M> AsyncInstrumentBuilder<'a, I, M>
where I: TryFrom<Self, Error = MetricsError> + AsyncInstrument<M>,

pub fn with_description<S: Into<Cow<'static, str>>>( self, description: S ) -> Self

Set the description for this instrument

pub fn with_unit(self, unit: Unit) -> Self

Set the unit for this instrument.

Unit is case sensitive(kb is not the same as kB).

Unit must be:

ASCII string

No longer than 63 characters

pub fn with_callback<F>(self, callback: F) -> Self
where F: Fn(&dyn AsyncInstrument<M>) + Send + Sync + 'static,

Set the callback to be called for this instrument.

pub fn try_init(self) -> Result<I>

Validate the instrument configuration and creates a new instrument.

pub fn init(self) -> I

Creates a new instrument.

Validate the instrument configuration and creates a new instrument.

Panics

Panics if the instrument cannot be created. Use try_init if you want to handle errors.