opentelemetry

Module propagation

OpenTelemetry Propagator interface

Cross-cutting concerns send their state to the next process using Propagators, which are defined as objects used to read and write context data to and from messages exchanged by the applications.

Propagators leverage the Context to inject and extract data for each cross-cutting concern, such as TraceContext and Baggage.

The Propagators API is expected to be leveraged by users writing instrumentation libraries.

Currently, the following Propagator types are supported:

TextMapPropagator, inject values into and extracts values from carriers as string key/value pairs

A binary Propagator type will be added in the future, See tracking issues).

Propagators uses Injector and Extractor to read and write context data to and from messages. Each specific Propagator type defines its expected carrier type, such as a string map or a byte array.

Re-exports

pub use composite::TextMapCompositePropagator;

pub use text_map_propagator::TextMapPropagator;

Modules

composite

Composite Propagator

text_map_propagator

TextMapPropagator

Traits

Extractor

Extractor provides an interface for removing fields from an underlying struct like HashMap

Injector

Injector provides an interface for adding fields from an underlying struct like HashMap