Module opentelemetry::propagation::text_map_propagator

Text Propagator

TextMapPropagator is a formatter to serialize and deserialize a value into a text format.

Structs

FieldIter

An iterator over fields of a TextMapPropagator

Traits

TextMapPropagator

Methods to inject and extract a value as text into injectors and extractors that travel in-band across process boundaries.