Trait opentelemetry::runtime::Runtime

pub trait Runtime: Clone + Send + Sync + 'static {
    type Interval: Stream + Send;
    type Delay: Future + Send;

    // Required methods
    fn interval(&self, duration: Duration) -> Self::Interval;
    fn spawn(&self, future: BoxFuture<'static, ()>);
    fn delay(&self, duration: Duration) -> Self::Delay;
}

A runtime is an abstraction of an async runtime like Tokio or async-std. It allows OpenTelemetry to work with any current and hopefully future runtime implementation.

Required Associated Types

type Interval: Stream + Send

A future stream, which returns items in a previously specified interval. The item type is not important.

type Delay: Future + Send

A future, which resolves after a previously specified amount of time. The output type is not important.

Required Methods

fn interval(&self, duration: Duration) -> Self::Interval

Create a Stream, which returns a new item every Duration.

fn spawn(&self, future: BoxFuture<'static, ()>)

Spawn a new task or thread, which executes the given future.

Note

This is mainly used to run batch span processing in the background. Note, that the function does not return a handle. OpenTelemetry will use a different way to wait for the future to finish when TracerProvider gets shutdown. At the moment this happens by blocking the current thread. This means runtime implementations need to make sure they can still execute the given future even if the main thread is blocked.

fn delay(&self, duration: Duration) -> Self::Delay

Return a new future, which resolves after the specified Duration.

Object Safety

This trait is not object safe.