Module opentelemetry::sdk::export::metrics

Available on crate feature metrics only.

Metrics Export

Modules

aggregation

Metrics aggregation

Structs

Accumulation

A container for the exported data for a single metric instrument and attribute set, as prepared by an Accumulator for the Processor.

ExportLine

Individually exported metric

ExportNumeric

A number exported as debug for serialization

Metadata

A container for the common elements for exported metric data that are shared by the Accumulator->Processor and Processor->Exporter steps.

Record

A container for the exported data for a single metric instrument and attribute set, as prepared by the Processor for the Exporter. This includes the effective start and end time for the aggregation.

StdoutExporter

An OpenTelemetry metric exporter that transmits telemetry to the local STDOUT or via the registered implementation of Write.

StdoutExporterBuilder

Configuration for a given stdout exporter.

Traits

AggregatorSelector

AggregatorSelector supports selecting the kind of Aggregator to use at runtime for a specific metric instrument.

AsDynProcessor

A utility extension to allow upcasting.

Checkpointer

The interface used to create checkpoints.

CheckpointerFactory

An interface for producing configured Checkpointer instances.

InstrumentationLibraryReader

InstrumentationLibraryReader is an interface for exporters to iterate over one instrumentation library of metric data at a time.

LockedCheckpointer

The interface used by a controller to coordinate the processor with accumulator(s) and exporter(s).

LockedProcessor

Metric data processor.

MetricsExporter

Exporter handles presentation of the checkpoint of aggregate metrics. This is the final stage of a metrics export pipeline, where metric data are formatted for a specific system.

Processor

Processor is responsible for deciding which kind of aggregation to use (via aggregation_selector), gathering exported results from the SDK during collection, and deciding over which dimensions to group the exported data.

Reader

Reader allows a controller to access a complete checkpoint of aggregated metrics from the Processor for a single library of metric data. This is passed to the Exporter which may then use ForEach to iterate over the collection of aggregated metrics.

Functions

accumulation

Allows Accumulator implementations to construct new Accumulations to send to Processors. The Descriptor, Attributes, Resource, and Aggregator represent aggregate metric events received over a single collection period.

record

Allows Processor implementations to construct export records. The Descriptor, Attributes, and Aggregator represent aggregate metric events received over a single collection period.

stdout

Create a new stdout exporter builder with the configuration for a stdout exporter.