Module opentelemetry::sdk::metrics::aggregators

Metric Aggregators

Structs

ArrayAggregator

An aggregator which stores metrics in an array.

DdSketchAggregator

DDSKetch quantile sketch algorithm

DdSketchConfig

DDSKetch Configuration.

HistogramAggregator

This aggregator observes events and counts them in pre-determined buckets. It also calculates the sum and count of all events.

LastValueAggregator

Aggregates last value events.

MinMaxSumCountAggregator

An Aggregator that aggregates events that form a distribution, keeping only the min, max, sum, and count.

SumAggregator

An aggregator for counter events.

Functions

array

Create a new default ArrayAggregator

ddsketch

An aggregator to calculate quantile

histogram

Create a new histogram for the given descriptor with the given boundaries

last_value

Create a new LastValueAggregator

min_max_sum_count

Create a new MinMaxSumCountAggregator

range_test

RangeTest is a common routine for testing for valid input values. This rejects NaN values. This rejects negative values when the metric instrument does not support negative values, including monotonic counter metrics and absolute ValueRecorder metrics.

sum

Create a new sum aggregator.