Struct opentelemetry::sdk::propagation::TextMapCompositePropagator

pub struct TextMapCompositePropagator { /* private fields */ }

Composite propagator

A propagator that chains multiple TextMapPropagator propagators together, injecting or extracting by their respective HTTP header names.

Injection and extraction from this propagator will preserve the order of the injectors and extractors passed in during initialization.

Examples

use opentelemetry::{
    baggage::BaggageExt,
    propagation::TextMapPropagator,
    trace::{TraceContextExt, Tracer, TracerProvider},
    Context, KeyValue,
};
use opentelemetry::sdk::propagation::{
    BaggagePropagator, TextMapCompositePropagator, TraceContextPropagator,
};
use opentelemetry::sdk::trace as sdktrace;
use std::collections::HashMap;

// First create 1 or more propagators
let baggage_propagator = BaggagePropagator::new();
let trace_context_propagator = TraceContextPropagator::new();

// Then create a composite propagator
let composite_propagator = TextMapCompositePropagator::new(vec![
    Box::new(baggage_propagator),
    Box::new(trace_context_propagator),
]);

// Then for a given implementation of `Injector`
let mut injector = HashMap::new();

// And a given span
let example_span = sdktrace::TracerProvider::default()
    .tracer("example-component")
    .start("span-name");

// with the current context, call inject to add the headers
composite_propagator.inject_context(
    &Context::current_with_span(example_span)
        .with_baggage(vec![KeyValue::new("test", "example")]),
    &mut injector,
);

// The injector now has both `baggage` and `traceparent` headers
assert!(injector.get("baggage").is_some());
assert!(injector.get("traceparent").is_some());

Implementations

impl TextMapCompositePropagator

pub fn new(propagators: Vec<Box<dyn TextMapPropagator + Send + Sync>>) -> Self

Constructs a new propagator out of instances of TextMapPropagator.