Trait opentelemetry::sdk::trace::ShouldSample

pub trait ShouldSample: Send + Sync + Debug {
    fn should_sample(
        &self, 
        parent_context: Option<&Context>, 
        trace_id: TraceId, 
        name: &str, 
        span_kind: &SpanKind, 
        attributes: &[KeyValue], 
        links: &[Link]
    ) -> SamplingResult;
}

The ShouldSample interface allows implementations to provide samplers which will return a sampling SamplingResult based on information that is typically available just before the Span was created.

Required methods

fn should_sample(
&self,
parent_context: Option<&Context>,
trace_id: TraceId,
name: &str,
span_kind: &SpanKind,
attributes: &[KeyValue],
links: &[Link]
) -> SamplingResult

Returns the SamplingDecision for a Span to be created.

Loading content...