Trait opentelemetry::sdk::trace::TraceRuntime

pub trait TraceRuntime: Runtime {
    type Receiver: Stream<Item = BatchMessage> + Send;
    type Sender: TrySend + Debug;

    // Required method
    fn batch_message_channel(
        &self,
        capacity: usize
    ) -> (Self::Sender, Self::Receiver);
}

Trace runtime is an extension to Runtime. Currently it provides a channel that used by BatchSpanProcessor.

Required Associated Types

type Receiver: Stream<Item = BatchMessage> + Send

A future stream to receive the batch messages from channels.

type Sender: TrySend + Debug

A batch messages sender that could be sent across thread safely.

Required Methods

fn batch_message_channel( &self, capacity: usize ) -> (Self::Sender, Self::Receiver)

Return the sender and receiver used to send batch message between tasks.