opentelemetry

Module trace

API for tracing applications and libraries.

The trace module includes types for tracking the progression of a single request while it is handled by services that make up an application. A trace is a tree of Spans which are objects that represent the work being done by individual services or components involved in a request as it flows through a system. This module implements the OpenTelemetry trace specification.

Getting Started

In application code:

use opentelemetry::trace::{Tracer, noop::NoopTracerProvider};
use opentelemetry::global;

fn init_tracer() {
    // Swap this no-op provider for your tracing service of choice (jaeger, zipkin, etc)
    let provider = NoopTracerProvider::new();

    // Configure the global `TracerProvider` singleton when your app starts
    // (there is a no-op default if this is not set by your application)
    let _ = global::set_tracer_provider(provider);
}

fn do_something_tracked() {
    // Then you can get a named tracer instance anywhere in your codebase.
    let tracer = global::tracer("my-component");

    tracer.in_span("doing_work", |cx| {
        // Traced app logic here...
    });
}

// in main or other app start
init_tracer();
do_something_tracked();

In library code:

use opentelemetry::{global, trace::{Span, Tracer, TracerProvider}};
use opentelemetry::InstrumentationScope;
use std::sync::Arc;

fn my_library_function() {
    // Use the global tracer provider to get access to the user-specified
    // tracer configuration
    let tracer_provider = global::tracer_provider();

    // Get a tracer for this library
    let scope = InstrumentationScope::builder("my_name")
        .with_version(env!("CARGO_PKG_VERSION"))
        .with_schema_url("https://opentelemetry.io/schemas/1.17.0")
        .build();

    let tracer = tracer_provider.tracer_with_scope(scope);

    // Create spans
    let mut span = tracer.start("doing_work");

    // Do work...

    // End the span
    span.end();
}

Overview

The tracing API consists of a three main traits:

TracerProviders are the entry point of the API. They provide access to Tracers.

Tracers are types responsible for creating Spans.

Spans provide the API to trace an operation.

Working with Async Runtimes

Exporting spans often involves sending data over a network or performing other I/O tasks. OpenTelemetry allows you to schedule these tasks using whichever runtime you are already using such as Tokio. When using an async runtime it’s best to use the batch span processor where the spans will be sent in batches as opposed to being sent once ended, which often ends up being more efficient.

Managing Active Spans

Spans can be marked as “active” for a given Context, and all newly created spans will automatically be children of the currently active span.

The active span for a given thread can be managed via get_active_span and mark_span_as_active.

use opentelemetry::{global, trace::{self, Span, Status, Tracer, TracerProvider}};

fn may_error(rand: f32) {
    if rand < 0.5 {
        // Get the currently active span to record additional attributes,
        // status, etc.
        trace::get_active_span(|span| {
            span.set_status(Status::error("value too small"));
        });
    }
}

// Get a tracer
let tracer = global::tracer("my_tracer");

// Create a span
let span = tracer.start("parent_span");

// Mark the span as active
let active = trace::mark_span_as_active(span);

// Any span created here will be a child of `parent_span`...

// Drop the guard and the span will no longer be active
drop(active)

Additionally Tracer::in_span can be used as shorthand to simplify managing the parent context.

use opentelemetry::{global, trace::Tracer};

// Get a tracer
let tracer = global::tracer("my_tracer");

// Use `in_span` to create a new span and mark it as the parent, dropping it
// at the end of the block.
tracer.in_span("parent_span", |cx| {
    // spans created here will be children of `parent_span`
});

Async active spans

Async spans can be propagated with TraceContextExt and FutureExt.

use opentelemetry::{Context, global, trace::{FutureExt, TraceContextExt, Tracer}};

async fn some_work() { }

// Get a tracer
let tracer = global::tracer("my_tracer");

// Start a span
let span = tracer.start("my_span");

// Perform some async work with this span as the currently active parent.
some_work().with_context(Context::current_with_span(span)).await;

Re-exports

pub use crate::SpanId;

pub use crate::TraceFlags;

pub use crate::TraceId;

Modules

noop

No-op trace implementation

Structs

Event

Events record things that happened during a Span’s lifetime.

Link

Link is the relationship between two Spans.

SamplingResult

The result of sampling logic for a given span.

SpanBuilder

SpanBuilder allows span attributes to be configured before the span has started.

SpanContext

Immutable portion of a Span which can be serialized and propagated.

SpanRef

A reference to the currently active span in this context.

TraceState

TraceState carries system-specific configuration data, represented as a list of key-value pairs. TraceState allows multiple tracing systems to participate in the same trace.

WithContext

A future, stream, or sink that has an associated context.

Enums

SamplingDecision

Decision about whether or not to sample

SpanKind

SpanKind describes the relationship between the Span, its parents, and its children in a trace.

Status

The status of a Span.

Traits

FutureExt

Extension trait allowing futures, streams, and sinks to be traced with a span.

Span

The interface for a single operation within a trace.

TraceContextExt

Methods for storing and retrieving trace data in a Context.

Tracer

The interface for constructing Spans.

TracerProvider

Types that can create instances of Tracer.

Functions

get_active_span

Executes a closure with a reference to this thread’s current span.

mark_span_as_active

Mark a given Span as active.